less differences between 32 and 64, and everywhere else where sizeof(size_t) is different

+++ include/daScript/simulate/simulate.h
@@ -155,7 +155,7 @@ namespace das
     struct SimVisitor {
         virtual void preVisit ( SimNode * ) { }
         virtual void cr () {}
-        virtual void op ( const char * /* name */, size_t /* sz */ = 0, const string & /* TT */ = string() ) {}
+        virtual void op ( const char * /* name */, uint32_t /* sz */ = 0, const string & /* TT */ = string() ) {}
         virtual void sp ( uint32_t /* stackTop */,  const char * /* op */ = "#sp" ) { }
         virtual void arg ( int32_t /* argV */,  const char * /* argN */  ) { }
         virtual void arg ( uint32_t /* argV */,  const char * /* argN */  ) { }
